reorganize file structure, get fields to populate on query
